requirements.txt
aniso8601==7.0.0
astroid==2.9.3
colorama==0.4.4
Django==2.2.13
django-cors-headers==3.7.0
django-filter==2.3.0
django-model-utils==4.0.0
djangorestframework==3.13.1
graphene==2.1.8
graphene-django==2.10.1
graphql-core==2.3.2
graphql-relay==2.0.1
importlib-metadata==4.11.3
isort==5.10.1
lazy-object-proxy==1.7.1
Markdown==3.3.7
mccabe==0.6.1
platformdirs==2.4.1
promise==2.3
psycopg2==2.9.3
pylint==2.12.2
python-dotenv==0.13.0
pytz==2020.1
Rx==1.6.1
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.3.1
toml==0.10.2
typing_extensions==4.0.1
wrapt==1.13.3
zipp==3.8.0
